requirements.txt
openai-agents
pydantic
httpx
rapidfuzz
PyMuPDF
python-magic
rich




